feat(web): add preview for digest delay and custom and refactor some code
