+++ apps/web/src/components/workflow/preview/chat/index.ts
@@ -1,1 +1,2 @@
 export * from './ChatPreview';
+export { ChatBasePreview } from './ChatBasePreview';
