+++ apps/web/src/components/workflow/preview/in-app/index.ts
@@ -1,1 +1,2 @@
 export * from './InAppPreview';
+export { InAppBasePreview } from './InAppBasePreview';
