+++ apps/web/src/components/workflow/preview/push/index.ts
@@ -1,1 +1,2 @@
 export * from './PushPreview';
+export * from './PushBasePreview';
