+++ apps/web/src/components/workflow/preview/sms/SmsPreview.tsx
@@ -1,9 +1,7 @@
-import styled from '@emotion/styled';
-import { colors } from '@novu/design-system';
 import { api } from '../../../../api';
 import { useEnvController } from '../../../../hooks/useEnvController';
 import { useMutation } from '@tanstack/react-query';
-import { MouseEventHandler, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useFormContext } from 'react-hook-form';
 import { useLocation } from 'react-router-dom';
 import { IForm } from '../../../../pages/templates/components/formTypes';
@@ -12,26 +10,7 @@ import { useNavigateToStepEditor } from '../../../../pages/templates/hooks/useNa
 import { usePreviewSmsTemplate } from '../../../../pages/templates/hooks/usePreviewSmsTemplate';
 import { useStepFormPath } from '../../../../pages/templates/hooks/useStepFormPath';
 import { useTemplateLocales } from '../../../../pages/templates/hooks/useTemplateLocales';
-import { LocaleSelect, MobileSimulator } from '../common';
-import { SmsBubble } from './SmsBubble';
-import { ErrorPrettyRender } from '../ErrorPrettyRender';
-
-const BodyContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  margin: auto 1.25rem 2.5rem 1.25rem;
-`;
-
-const LocaleSelectStyled = styled(LocaleSelect)`
-  .mantine-Select-input {
-    color: ${({ theme }) => (theme.colorScheme === 'dark' ? colors.white : colors.B60)};
-  }
-
-  .mantine-Input-rightSection {
-    color: ${({ theme }) => (theme.colorScheme === 'dark' ? colors.white : colors.B60)} !important;
-  }
-`;
+import { SmsBasePreview } from './SmsBasePreview';
 
 export const SmsPreview = ({
   showPreviewAsLoading = false,
@@ -79,7 +58,7 @@ export const SmsPreview = ({
   );
 
   return (
-    <SmsPreviewComponent
+    <SmsBasePreview
       content={bridge ? bridgeContent : previewContent}
       onLocaleChange={onLocaleChange}
       locales={locales}
@@ -90,56 +69,5 @@ export const SmsPreview = ({
       error={bridge ? undefined : templateError}
       previewError={previewError}
     />
-  );
-};
-
-export const SmsPreviewComponent = ({
-  content,
-  loading = false,
-  error,
-  showEditOverlay = false,
-  onLocaleChange,
-  previewError,
-  selectedLocale,
-  locales,
-  onEditClick,
-}: {
-  content: string;
-  loading?: boolean;
-  error?: string;
-  previewError?: any;
-  showEditOverlay?: boolean;
-  onLocaleChange: (locale: string) => void;
-  selectedLocale?: string;
-  locales: any[];
-  onEditClick?: MouseEventHandler<HTMLButtonElement>;
-}) => {
-  if (previewError) {
-    return (
-      <div style={{ marginTop: 20, padding: 10 }}>
-        <ErrorPrettyRender error={previewError} />
-      </div>
-    );
-  }
-
-  return (
-    <MobileSimulator withBackground={false}>
-      <BodyContainer>
-        <LocaleSelectStyled
-          isLoading={loading}
-          locales={locales}
-          value={selectedLocale}
-          onLocaleChange={onLocaleChange}
-          dropdownPosition="top"
-        />
-        <SmsBubble
-          onEditClick={onEditClick}
-          isLoading={loading}
-          text={content}
-          error={error}
-          withOverlay={showEditOverlay}
-        />
-      </BodyContainer>
-    </MobileSimulator>
   );
 };
