+++ apps/web/src/components/workflow/preview/sms/index.ts
@@ -1,1 +1,2 @@
 export * from './SmsPreview';
+export * from './SmsBasePreview';
